requirements.txt
fastapi
uvicorn
python-multipart
PyPDF2
python-docx
markdown
httpx
pydantic
pydantic-settings
python-dotenv
exa-py
google-genai
pinecone

